refactor: AppError 및 AppException 위치 변경 (core:model -> core:common)
* **refactor: 에러 관련 클래스 패키지 이동**
    * 공통으로 사용되는 `AppError` 및 `AppException` 클래스를 `core:model` 모듈에서 `core:common` 모듈의 `com.team.prezel.core.common.error` 패키지로 이동했습니다.

* **build: 모듈 의존성 설정 업데이트**
    * `AndroidFeatureImplConventionPlugin`에 `core-common` 의존성을 추가하여 Feature 모듈에서 사용할 수 있도록 구성했습니다.
    * `core:domain` 모듈의 `build.gradle.kts`에 `core-common` 프로젝트 의존성을 추가했습니다.

* **refactor: 변경된 에러 클래스 참조 수정 및 코드 정리**
    * `CheckLoginStatusUseCase`, `TermsViewModel`, `AppErrorExt` 등 전반적인 모듈에서 변경된 패키지 경로에 맞춰 import 문을 업데이트했습니다.
    * `TermsViewModel` 및 `AppErrorExt` 내 `when` 식의 들여쓰기를 수정하여 가독성을 개선했습니다.
