+++ Prezel/build-logic/convention/src/main/java/com/team/prezel/buildlogic/convention/plugin/AndroidFeatureImplConventionPlugin.kt
@@ -23,6 +23,7 @@ class AndroidFeatureImplConventionPlugin : Plugin<Project> {
                 "implementation"(project(":core-ui"))
                 "implementation"(project(":core-designsystem"))
                 "implementation"(project(":core-navigation"))
+                "implementation"(project(":core-common"))
                 "implementation"(libs.findLibrary("androidx.navigation3.ui").get())
                 "implementation"(libs.findLibrary("androidx.hilt.lifecycle.viewmodel.compose").get())
                 "implementation"(libs.findLibrary("timber").get())
