+++ Prezel/core/common/src/main/kotlin/com/team/prezel/core/common/error/AppError.kt
@@ -1,4 +1,4 @@
-package com.team.prezel.core.model.base
+package com.team.prezel.core.common.error
 
 enum class AppError {
     UNAUTHORIZED,
