+++ Prezel/core/common/src/main/kotlin/com/team/prezel/core/common/error/AppException.kt
@@ -1,4 +1,4 @@
-package com.team.prezel.core.model.base
+package com.team.prezel.core.common.error
 
 class AppException(
     val error: AppError,
